fix(i18n): serve all 8 advertised locales, dedupe landing page, repair /admin & /audit-log routing
The proxy, sitemap, hreflang alternates and language switcher all
advertised zh/ko/es/pt, but no routes existed — visitors with those
Accept-Language values (or switcher picks) were redirected to /es etc.
and got a 404, and the sitemap submitted dead URLs to crawlers.

- Extract the en landing page (newest design) into a shared
  <LandingPage lang> server component; en/de/fr become 3-line wrappers
  and es/pt/zh/ko routes are added (dictionaries already existed).
  de/fr pick up the dict-driven sections they had drifted behind on.
  /ja keeps its deliberately Japan-specific page. Net: −3,400 LOC of
  copy-pasted JSX, one source of truth for 7 locales.
- Shared landingMetadata() helper for canonical + hreflang in every
  locale layout.
- Dictionaries: add missing ja socialProof.betaSubtitle key and
  translate strings left in English across de/fr/es/pt/zh/ko
  (features.viewAll, socialProof, agentResilience, nlayer, comparison).
- proxy.ts: /admin and /audit-log were missing from the locale-redirect
  skip list, so they were rewritten to /en/admin etc. and 404ed for
  everyone. Route lists are now built from single PUBLIC_PAGES /
  APP_ROUTES constants (also used for auth protection, which now covers
  /audit-log too). Verified live: / with Accept-Language es/ko → 200,
  /admin & /audit-log → login redirect, /ja/dashboard → /dashboard.
- breadcrumb: reuse locales from i18n config instead of a drifting
  hardcoded copy; navbar: drop a duplicated setMobileOpen effect.
