+++ src/app/es/layout.tsx
@@ -1,13 +1,13 @@
 import { landingMetadata } from "@/i18n/landing-metadata";
 
 export function generateMetadata() {
-  return landingMetadata("de");
+  return landingMetadata("es");
 }
 
-export default function DeLayout({
+export default function EsLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  return <div lang="de">{children}</div>; // I18N-04
+  return <div lang="es">{children}</div>; // I18N-04
 }
