+++ src/app/es/page.tsx
@@ -1,0 +1,5 @@
+import { LandingPage } from "@/components/landing-page";
+
+export default function EsHomePage() {
+  return <LandingPage lang="es" />;
+}
