+++ src/app/fr/layout.tsx
@@ -1,21 +1,7 @@
-import { getDictionary } from "@/i18n/get-dictionary";
-import type { Metadata } from "next";
+import { landingMetadata } from "@/i18n/landing-metadata";
 
-const BASE = "https://faultray.com";
-const LOCALES = ["en", "ja", "de", "fr", "zh", "ko", "es", "pt"] as const;
-
-export async function generateMetadata(): Promise<Metadata> {
-  const dict = await getDictionary("fr");
-  return {
-    title: dict.metadata.title,
-    description: dict.metadata.description,
-    alternates: {
-      canonical: `${BASE}/fr`,
-      languages: Object.fromEntries(
-        LOCALES.map((locale) => [locale, `${BASE}/${locale}`])
-      ) as Record<string, string>,
-    },
-  };
+export function generateMetadata() {
+  return landingMetadata("fr");
 }
 
 export default function FrLayout({
