+++ src/app/ko/page.tsx
@@ -1,0 +1,5 @@
+import { LandingPage } from "@/components/landing-page";
+
+export default function KoHomePage() {
+  return <LandingPage lang="ko" />;
+}
