+++ src/app/pt/layout.tsx
@@ -1,13 +1,13 @@
 import { landingMetadata } from "@/i18n/landing-metadata";
 
 export function generateMetadata() {
-  return landingMetadata("de");
+  return landingMetadata("pt");
 }
 
-export default function DeLayout({
+export default function PtLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  return <div lang="de">{children}</div>; // I18N-04
+  return <div lang="pt">{children}</div>; // I18N-04
 }
