+++ src/app/pt/page.tsx
@@ -1,0 +1,5 @@
+import { LandingPage } from "@/components/landing-page";
+
+export default function PtHomePage() {
+  return <LandingPage lang="pt" />;
+}
