+++ src/app/zh/page.tsx
@@ -1,0 +1,5 @@
+import { LandingPage } from "@/components/landing-page";
+
+export default function ZhHomePage() {
+  return <LandingPage lang="zh" />;
+}
