+++ src/i18n/dictionaries/es.json
@@ -127,7 +127,7 @@
         "desc": "Panel web completo con editor de topología, explorador de escenarios, drill-down N-Layer, mapa de calor, borradores de investigación DORA y resúmenes ejecutivos."
       }
     ],
-    "viewAll": "View Full Feature Details"
+    "viewAll": "Ver todas las funciones en detalle"
   },
   "agentResilience": {
     "badge": "Novedad en v11.0",
@@ -249,7 +249,7 @@
       999
     ],
     "taxNote": "",
-    "japaneseSupportBadge": "Japanese support included",
+    "japaneseSupportBadge": "Soporte en japonés incluido",
     "plans": [
       {
         "name": "Gratis",
@@ -269,11 +269,11 @@
     ]
   },
   "socialProof": {
-    "heading": "Trusted by the Developer Community",
+    "heading": "Con la confianza de la comunidad de desarrolladores",
     "githubStars": "GitHub Stars",
-    "betaMessage": "Currently in beta. We'd love your feedback.",
-    "betaSubtitle": "We are in early access. Your feedback helps us improve.",
-    "betaCta": "Send Feedback",
+    "betaMessage": "Actualmente en beta. Nos encantaría conocer tu opinión.",
+    "betaSubtitle": "Estamos en acceso anticipado. Tus comentarios nos ayudan a mejorar.",
+    "betaCta": "Enviar comentarios",
     "openSource": "Open Source"
   },
   "roi": {
@@ -296,7 +296,7 @@
     "resources": "Recursos",
     "legal": "Legal",
     "copyright": "© 2026 FaultRay. Código abierto bajo Apache License 2.0.",
-    "japaneseSupport": "Japanese support available"
+    "japaneseSupport": "Soporte en japonés disponible"
   },
   "languageSwitcher": {
     "label": "Idioma"
