+++ src/i18n/dictionaries/ko.json
@@ -127,7 +127,7 @@
         "desc": "토폴로지 에디터, 시나리오 탐색기, N-Layer 드릴다운, 히트맵, DORA 연구 초안, 경영 요약을 갖춘 완전한 웹 대시보드."
       }
     ],
-    "viewAll": "View Full Feature Details"
+    "viewAll": "전체 기능 자세히 보기"
   },
   "agentResilience": {
     "badge": "v11.0 신기능",
@@ -249,7 +249,7 @@
       999
     ],
     "taxNote": "",
-    "japaneseSupportBadge": "Japanese support included",
+    "japaneseSupportBadge": "일본어 지원 포함",
     "plans": [
       {
         "name": "무료",
@@ -269,11 +269,11 @@
     ]
   },
   "socialProof": {
-    "heading": "Trusted by the Developer Community",
+    "heading": "개발자 커뮤니티가 신뢰하는 도구",
     "githubStars": "GitHub Stars",
-    "betaMessage": "Currently in beta. We'd love your feedback.",
-    "betaSubtitle": "We are in early access. Your feedback helps us improve.",
-    "betaCta": "Send Feedback",
+    "betaMessage": "현재 베타 버전입니다. 피드백을 보내주세요.",
+    "betaSubtitle": "얼리 액세스 단계입니다. 보내주신 피드백이 개선에 큰 도움이 됩니다.",
+    "betaCta": "피드백 보내기",
     "openSource": "Open Source"
   },
   "roi": {
@@ -296,7 +296,7 @@
     "resources": "리소스",
     "legal": "법적 고지",
     "copyright": "© 2026 FaultRay. Apache 2.0Apache 2.0 라이선스 하에 오픈소스.",
-    "japaneseSupport": "Japanese support available"
+    "japaneseSupport": "일본어 지원 제공"
   },
   "languageSwitcher": {
     "label": "언어"
