+++ src/i18n/dictionaries/zh.json
@@ -127,7 +127,7 @@
         "desc": "完整的 Web 仪表盘，包含拓扑编辑器、场景浏览器、N-Layer 钻取、热力图、DORA 研究草稿和高管摘要。"
       }
     ],
-    "viewAll": "View Full Feature Details"
+    "viewAll": "查看全部功能详情"
   },
   "agentResilience": {
     "badge": "v11.0 新功能",
@@ -249,7 +249,7 @@
       999
     ],
     "taxNote": "",
-    "japaneseSupportBadge": "Japanese support included",
+    "japaneseSupportBadge": "含日语支持",
     "plans": [
       {
         "name": "免费版",
@@ -269,11 +269,11 @@
     ]
   },
   "socialProof": {
-    "heading": "Trusted by the Developer Community",
+    "heading": "深受开发者社区信赖",
     "githubStars": "GitHub Stars",
-    "betaMessage": "Currently in beta. We'd love your feedback.",
-    "betaSubtitle": "We are in early access. Your feedback helps us improve.",
-    "betaCta": "Send Feedback",
+    "betaMessage": "目前处于测试阶段，期待您的反馈。",
+    "betaSubtitle": "我们正处于抢先体验阶段，您的反馈将帮助我们不断改进。",
+    "betaCta": "发送反馈",
     "openSource": "Open Source"
   },
   "roi": {
@@ -296,7 +296,7 @@
     "resources": "资源",
     "legal": "法律",
     "copyright": "© 2026 FaultRay. 基于 Apache 2.0（商业源代码许可证）开源。",
-    "japaneseSupport": "Japanese support available"
+    "japaneseSupport": "提供日语支持"
   },
   "languageSwitcher": {
     "label": "语言"
